+++ packages/kit/src/core/context.ts
@@ -1,0 +1,19 @@
+import {createContext} from "react"
+import {FlowNetwork} from "./types"
+
+export type FlowConfig = {
+  accessNodeUrl?: string
+  appDetailTitle?: string
+  appDetailIcon?: string
+  appDetailDescription?: string
+  appDetailUrl?: string
+  discoveryWallet?: string
+  discoveryWalletMethod?: string
+  fclLimit?: number
+  flowNetwork?: FlowNetwork
+  serviceOpenIdScopes?: string[]
+  walletconnectProjectId?: string
+  walletconnectDisableNotifications?: boolean
+}
+
+export const FlowConfigContext = createContext<FlowConfig>({})