+++ packages/fcl-rainbowkit-adapter/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @onflow/fcl-rainbowkit-adapter
 
-Initial release of the FCL Rainbowkit Adapter.  This adapter allows you to use FCL wallets within Rainbowkit in a shared session with FCL-JS.
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @onflow/fcl@1.16.1
+  - @onflow/fcl-ethereum-provider@0.0.3
+  - @onflow/fcl-wagmi-adapter@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`6c4d6c3df669883e34951bc1d26edb95441f124d`](https://github.com/onflow/fcl-js/commit/6c4d6c3df669883e34951bc1d26edb95441f124d)]:
+  - @onflow/fcl@1.16.0
+  - @onflow/fcl-ethereum-provider@0.0.2
+  - @onflow/fcl-wagmi-adapter@0.0.2
+
+## 0.0.1
+
+Initial release of the FCL Rainbowkit Adapter. This adapter allows you to use FCL wallets within Rainbowkit in a shared session with FCL-JS.
 
 Read more about this package in [FLIP-316](https://github.com/onflow/flips/pull/317) and the official [Flow Developer Documentation](https://developers.flow.com/tools/clients/fcl-js/cross-vm/rainbowkit-adapter).